requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
exceptiongroup==1.1.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
outcome==1.2.0
psycopg2-binary==2.9.5
PySocks==1.7.1
pytz==2022.6
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
yarl==1.8.2
